requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
overpy
requests
python-dotenv
